requirements.txt
click==8.1.3
cmake==3.18.2
colorama==0.4.6
numpy==1.24.3
opencv-python==4.7.0.72
pandas==2.0.1
Pillow==9.5.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqt5-tools==5.15.9.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
six==1.16.0
tzdata==2023.3
face-recognition-models==0.3.0
